Split ApiClient into own file
